fix: move components and FAQ page to own folders
