+++ src/components/Accordions/GeneralAccordion/GeneralAccordion.jsx
