+++ src/components/Accordions/RegisterAccordion/RegisterAccordion.jsx
