+++ src/components/Accordions/TechnicalAccordion/TechnicalAccordion.jsx
